# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/households/{householdId}/uploads": {
      "post": {
        "operationId": "UploadsController_upload",
        "parameters": [],
        "responses": {
          "201": {
            "description": "URL publique de l'image téléversée."
          },
          "400": {
            "description": "Les données envoyées sont invalides."
          },
          "401": {
            "description": "L'utilisateur n'est pas authentifié, ou le token est invalide ou expiré."
          },
          "403": {
            "description": "L'utilisateur n'a pas les droits nécessaires pour effectuer cette action."
          },
          "404": {
            "description": "La ressource demandée n'existe pas, ou n'appartient pas à ce household."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Téléverse une image de couverture (JPEG/PNG/WebP, validé par signature binaire, 10 Mo max).",
        "tags": [
          "uploads"
        ]
      }
    },
    "/households/{householdId}/uploads/{uploadId}": {
      "delete": {
        "operationId": "UploadsController_remove",
        "parameters": [
          {
            "name": "uploadId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Fichier supprimé."
          },
          "401": {
            "description": "L'utilisateur n'est pas authentifié, ou le token est invalide ou expiré."
          },
          "403": {
            "description": "L'utilisateur n'a pas les droits nécessaires pour effectuer cette action."
          },
          "404": {
            "description": "La ressource demandée n'existe pas, ou n'appartient pas à ce household."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Supprime un fichier précédemment téléversé.",
        "tags": [
          "uploads"
        ]
      }
    },
...
  },
  "info": {
    "title": "Notre Nid API",
    "description": "API de gestion de collection partagée pour Notre Nid.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}